move GroupedData to tsdata to share between nc and md
